backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
psycopg[pool]==3.1.18
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
email-validator==2.3.0
dnspython==2.7.0
databricks-sdk==0.23.0
psycopg-binary==3.1.18
openpyxl==3.1.2
pandas==2.2.0
Faker==24.11.0